# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hydrogen Laravel Generator APIs",
        "contact": {
            "email": "amrmohamed257@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/cart-items?include=cart,product": {
            "get": {
                "tags": [
                    "CartItem"
                ],
                "summary": "getCartItemList",
                "description": "Get all CartItem",
                "operationId": "02abe30b799e0053531eb35496cfe30c",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "cartItems": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CartItem"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart-items": {
            "post": {
                "tags": [
                    "CartItem"
                ],
                "summary": "createCartItem",
                "description": "Create CartItem",
                "operationId": "0594d84a244ba4a9806032ec59ba1298",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "cartItem": {
                                            "$ref": "#/components/schemas/CartItem"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart-items/{id}": {
            "get": {
                "tags": [
                    "CartItem"
                ],
                "summary": "getCartItemItem",
                "description": "Get CartItem",
                "operationId": "3d3bd745d1de6ef57eff6a63987426f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of CartItem",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "cartItem": {
                                            "$ref": "#/components/schemas/CartItem"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CartItem"
                ],
                "summary": "updateCartItem",
                "description": "Update CartItem",
                "operationId": "ec4bd06028ad23246255ff6683face1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of CartItem",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "cartItem": {
                                            "$ref": "#/components/schemas/CartItem"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CartItem"
                ],
                "summary": "deleteCartItem",
                "description": "Delete CartItem",
                "operationId": "2cf9e15188ff7329d3ffa06862da4f85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of CartItem",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CartItem": {
                "required": [],
                "properties": {
                    "cart_id": {
                        "nullable": false,
                        "description": "Cart id",
                        "type": "integer",
                        "readOnly": false
                    },
                    "product_id": {
                        "nullable": false,
                        "description": "Product id",
                        "type": "integer",
                        "readOnly": false
                    },
                    "quantity": {
                        "nullable": false,
                        "description": "Quantity of items",
                        "type": "integer",
                        "readOnly": false
                    },
                    "created_at": {
                        "nullable": true,
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "deleted_at": {
                        "nullable": true,
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "nullable": true,
                        "schema": "CartItem",
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}